Finished 1.1.5
Made it so the program capitalizes the first letter of the first/last names and the location and makes everything else lowercase, added a print statement to display the name at the end, deleted test class. Also added everything to bluej because it looks better/is probably easier to turn in.