1/16 - There is a God.
Finally figured out how to create, save to, and read from files (at least for strings). App can now be closed and reopened and retain user's synthPass (saved to file "userSynthPass"). I'm in awe. Anyways, password can currently be written over by simply pressing "set your synth" and creating a new synth so this currently isn't in any way a secure system, but still this is def a step in the right direction.